RUN/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
audioread==3.0.1
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
duckduckgo_search==8.0.2
faiss-cpu==1.11.0
ffmpeg==1.4
filelock==3.18.0
fsspec==2025.5.0
gitdb==4.0.12
GitPython==3.1.44
groq==0.25.0
gTTS==2.5.4
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
lxml==5.4.0
MarkupSafe==3.0.2
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.41.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
primp==0.15.0
proglog==0.1.12
protobuf==6.31.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.5.24
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.3
triton==3.3.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
youtube-dl==2021.12.17
yt-dlp==2025.5.22
